# swagger.yaml
openapi: "3.0.3"
info:
  title: MySlotMate API
  description: |
    Backend API for **MySlotMate** – a slot-booking platform with wallet-based payments,
    Razorpay payouts, Aadhar (Setu OKYC) verification, AWS S3 file uploads,
    and real-time inbox messaging.

    Every response follows the standard envelope:
    ```json
    { "success": true|false, "data": ..., "message": "...", "error": "..." }
    ```
  version: "2.0.0"
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: "8080"
...
paths:
...
  /events/:
    post:
      tags: [Events]
      summary: Create a new event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreateRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /events/{eventID}:
    get:
      tags: [Events]
      summary: Get event details
      operationId: getEvent
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [Events]
      summary: Update an event
      operationId: updateEvent
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdateRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /events/host/{hostID}:
    get:
      tags: [Events]
      summary: List all events for a host
      operationId: getEventsByHost
      parameters:
        - $ref: "#/components/parameters/HostIDPath"
      responses:
        "200":
          description: Array of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_EventList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /events/host/{hostID}/filtered:
    get:
      tags: [Events]
      summary: Filtered event search for a host
      operationId: getHostEventsFiltered
      parameters:
        - $ref: "#/components/parameters/HostIDPath"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/EventStatus"
        - name: mood
          in: query
          schema:
            $ref: "#/components/schemas/EventMood"
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Filtered events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_EventList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /events/calendar/{hostID}:
    get:
      tags: [Events]
      summary: Calendar view of events for a host
      operationId: getCalendarEvents
      parameters:
        - $ref: "#/components/parameters/HostIDPath"
      responses:
        "200":
          description: Events for calendar rendering
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_EventList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /events/{eventID}/publish:
    post:
      tags: [Events]
      summary: Publish a draft event (status → live)
      operationId: publishEvent
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_id]
              properties:
                host_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Event published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /events/{eventID}/pause:
    post:
      tags: [Events]
      summary: Pause a live event (status → paused)
      operationId: pauseEvent
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_id]
              properties:
                host_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Event paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /events/{eventID}/resume:
    post:
      tags: [Events]
      summary: Resume a paused event (status → live)
      operationId: resumeEvent
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_id]
              properties:
                host_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Event resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /events/{eventID}/attendees:
    get:
      tags: [Events]
      summary: List confirmed attendees for an event
      operationId: getEventAttendees
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      responses:
        "200":
          description: Array of bookings (confirmed attendees)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_BookingList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Bookings ────────────────────────────────
...
  /reviews/event/{eventID}:
    get:
      tags: [Reviews]
      summary: List reviews for an event
      operationId: getReviewsByEvent
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      responses:
        "200":
          description: Array of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_ReviewList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /inbox/event/{eventID}:
    get:
      tags: [Inbox]
      summary: Get all messages for an event thread
      operationId: getEventMessages
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      responses:
        "200":
          description: Array of inbox messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_InboxMessageList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID token passed as `Authorization: Bearer <token>`. Admin endpoints additionally require the token's email to match the `ADMIN_EMAIL` env variable."

  # ── Parameters ──────────────────────────────
  parameters:
    HostIDPath:
      name: hostID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Host UUID
...
    EventIDPath:
      name: eventID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Event UUID
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ── Schemas ─────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string

    # ── Request Bodies ────────────────────────
...
    EventCreateRequest:
      type: object
      required: [host_id, title, time, capacity]
      properties:
        host_id:
          type: string
          format: uuid
        title:
          type: string
          example: "Weekend Football Slot"
        hook_line:
          type: string
          nullable: true
        mood:
          $ref: "#/components/schemas/EventMood"
        description:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        gallery_urls:
          type: array
          items:
            type: string
        time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        is_online:
          type: boolean
          default: false
        location:
          type: string
          nullable: true
        location_lat:
          type: number
          format: double
          nullable: true
        location_lng:
          type: number
          format: double
          nullable: true
        duration_minutes:
          type: integer
          nullable: true
        capacity:
          type: integer
          minimum: 1
          example: 20
        min_group_size:
          type: integer
          nullable: true
        max_group_size:
          type: integer
          nullable: true
        price_cents:
          type: integer
          format: int64
          nullable: true
          description: Per guest in paise; null = free
        is_free:
          type: boolean
          default: false
        is_recurring:
          type: boolean
          default: false
        recurrence_rule:
          type: string
          nullable: true
          description: "iCal rule, e.g. FREQ=WEEKLY;BYDAY=MO"
        cancellation_policy:
          $ref: "#/components/schemas/CancellationPolicy"
        ai_suggestion:
          type: string
          nullable: true
    EventUpdateRequest:
      type: object
      description: All fields are optional — only provided fields are updated.
      properties:
        host_id:
          type: string
          format: uuid
          description: Required for ownership verification
        title:
          type: string
          nullable: true
        hook_line:
          type: string
          nullable: true
        mood:
          $ref: "#/components/schemas/EventMood"
        description:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        gallery_urls:
          type: array
          items:
            type: string
        time:
          type: string
          format: date-time
          nullable: true
        end_time:
          type: string
          format: date-time
          nullable: true
        is_online:
          type: boolean
          nullable: true
        location:
          type: string
          nullable: true
        location_lat:
          type: number
          format: double
          nullable: true
        location_lng:
          type: number
          format: double
          nullable: true
        duration_minutes:
          type: integer
          nullable: true
        capacity:
          type: integer
          nullable: true
        min_group_size:
          type: integer
          nullable: true
        max_group_size:
          type: integer
          nullable: true
        price_cents:
          type: integer
          format: int64
          nullable: true
        is_free:
          type: boolean
          nullable: true
        is_recurring:
          type: boolean
          nullable: true
        recurrence_rule:
          type: string
          nullable: true
        cancellation_policy:
          $ref: "#/components/schemas/CancellationPolicy"
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        host_id:
          type: string
          format: uuid
        title:
          type: string
        hook_line:
          type: string
          nullable: true
        mood:
          $ref: "#/components/schemas/EventMood"
        description:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        gallery_urls:
          type: array
          items:
            type: string
        is_online:
          type: boolean
        location:
          type: string
          nullable: true
        location_lat:
          type: number
          format: double
          nullable: true
        location_lng:
          type: number
          format: double
          nullable: true
        duration_minutes:
          type: integer
          nullable: true
        min_group_size:
          type: integer
          nullable: true
        max_group_size:
          type: integer
          nullable: true
        capacity:
          type: integer
        price_cents:
          type: integer
          format: int64
          nullable: true
        is_free:
          type: boolean
        time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        is_recurring:
          type: boolean
        recurrence_rule:
          type: string
          nullable: true
        cancellation_policy:
          $ref: "#/components/schemas/CancellationPolicy"
        status:
          $ref: "#/components/schemas/EventStatus"
        published_at:
          type: string
          format: date-time
          nullable: true
        paused_at:
          type: string
          format: date-time
          nullable: true
        ai_suggestion:
          type: string
          nullable: true
        avg_rating:
          type: number
          format: double
          nullable: true
        total_bookings:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        quantity:
          type: integer
        status:
          $ref: "#/components/schemas/BookingStatus"
        payment_id:
          type: string
          format: uuid
          nullable: true
        idempotency_key:
          type: string
          nullable: true
        amount_cents:
          type: integer
          format: int64
          nullable: true
        service_fee_cents:
          type: integer
          format: int64
          nullable: true
        net_earning_cents:
          type: integer
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        cancelled_at:
          type: string
          format: date-time
          nullable: true
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        name:
          type: string
          nullable: true
        description:
          type: string
        photo_urls:
          type: array
          items:
            type: string
        reply:
          type: array
          items:
            type: string
        sentiment_score:
          type: number
          format: double
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InboxMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        sender_type:
          $ref: "#/components/schemas/MessageSenderType"
        sender_id:
          type: string
          format: uuid
          nullable: true
        message:
          type: string
        attachment_url:
          type: string
          nullable: true
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    BookingStatus:
      type: string
      enum: [pending, confirmed, cancelled, refunded]
...
    EventStatus:
      type: string
      enum: [draft, live, paused]
    EventMood:
      type: string
      enum: [adventure, social, wellness, chill, romantic, intellectual, foodie, nightlife]
    CancellationPolicy:
      type: string
      enum: [flexible, moderate, strict]
...
    MessageSenderType:
      type: string
      enum: [system, host, guest]

    # ── Wrapped Success Responses ─────────────
...
    SuccessResponse_Event:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Event"
    SuccessResponse_EventList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
...
    SuccessResponse_BookingList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
...
    SuccessResponse_ReviewList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Review"
...
    SuccessResponse_InboxMessageList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/InboxMessage"
...